refactor: move articles query to shared file
